style(ruff): apply low-risk fixes for simplifications and redefinitions
17 fixes across categories:
- F811 redefined-while-unused (4) - removed duplicate imports
- RUF005 collection-literal-concatenation (5) - use [*foo, bar]
- RUF015 unnecessary-iterable-allocation-for-first-element (3) - next(gen)
- SIM105 suppressible-exception (1) - contextlib.suppress
- SIM103 needless-bool (1) - return condition directly
- SIM110 reimplemented-builtin (1) - any()
- SIM118 in-dict-keys (1)
- RUF046 unnecessary-cast-to-int (1)

Also moved misplaced `import contextlib` in draft/state.py to top of imports.

All 1218 tests pass.

+++ scripts/simulate_draft.py
@@ -371,7 +371,7 @@ def run_simulation(
             # Strategy noise: sometimes take the 2nd or 3rd rec instead.
             # Normal distribution: ~68% take #1, ~27% take #2, ~4% take #3.
             if strategy_noise > 0 and pick_name is not None:
-                skip = min(abs(int(round(rng.normal(0, strategy_noise)))), 4)  # cap at 5th-best
+                skip = min(abs(round(rng.normal(0, strategy_noise))), 4)  # cap at 5th-best
                 if skip > 0:
                     filled = get_filled_positions(
                         tracker.user_roster_ids,
